% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignmentAgreement().R
\name{AssignmentAgreement}
\alias{AssignmentAgreement}
\title{Get agreed assignments}
\usage{
AssignmentAgreement(SDT)
}
\arguments{
\item{SDT}{A sequence data table with both Idtaxa and Blast assignments output by SimpleMetaPipeline.}
}
\value{
An updated sequence data table contianing only agreeing assignments. Note that assignment quality metrics are stripped out by ths function, so any filtering based on these metrics should be done first.
}
\description{
This function tests for agreement between Idtaxa and Blast assignments and returns only those assignments where algorithms agree.
}
\examples{
AssignmentAgreement(SDT=SDT)
}
