% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMultiRarefiedRichnessEstimates().R
\name{GetMultiRarefiedRichnessEstimates}
\alias{GetMultiRarefiedRichnessEstimates}
\title{Get replicate richness estiamtes by rarefying multiple times.}
\usage{
GetMultiRarefiedRichnessEstimates(ps, RarefyDepth = 2000, NumReplicates = 100)
}
\arguments{
\item{ps}{A phyloseq object as input.}

\item{RarefyDepth}{The deth to rarefy to. All samples shallower than this depth are discarded.}

\item{NumReplicates}{The number of rarefied replicates to produce}
}
\value{
A tibble with mutliple richness metrics for all samples deeper than the rarefy depth. Each has all metadata from the phyloseq object along with the random seed used to generate it.
}
\description{
This function produces multi-rarefied richness estiamtes from a hyloseq object.
}
\examples{
GetMultiRarefiedRichnessEstimates(ps=ps, RarefyDepth=2000)
}
