% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\docType{data}
\name{ps}
\alias{ps}
\title{A small dummy phyloseq object}
\format{
\subsection{\code{phyloseq}}{

phyloseq-class experiment-level object:
\describe{
\item{otu_table}{110 taxa by 2 samples}
\item{sample_data}{2 samples by 1 variable}
\item{tax_table}{110 taxa by 7 ranks}
\item{refseq}{110 sequences}
...
}
}
}
\usage{
ps
}
\description{
A small dummy phyloseq object
}
\keyword{datasets}
