% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT.R
\docType{data}
\name{SDT}
\alias{SDT}
\title{A small dummy SequenceDataTable}
\format{
\subsection{\code{SDT}}{

a data frame with 37 columns and 331 rows
This is a standard sequence data table output by SmpleMetaPipeline. It contains all possible columns
i.e. it is the output if both Idtaxa and BLAST assignments are applied
...
}
}
\usage{
SDT
}
\description{
A small dummy SequenceDataTable
}
\keyword{datasets}
