% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqDataTable2Phyloseq().R
\name{SeqDataTable2Phyloseq}
\alias{SeqDataTable2Phyloseq}
\title{Load a Matrix}
\usage{
SeqDataTable2Phyloseq(
  SeqDataTablePath,
  clustering = "ESV",
  Metadata = NULL,
  assignment = "Idtaxa",
  BLASTThreshold = NULL,
  ClusterAssignment = "RepresentativeSequence",
  ReformatSampleNames = TRUE,
  HandleDuplicates = TRUE
)
}
\arguments{
\item{SeqDataTablePath}{Path to the input SeqDataTable.}

\item{clustering}{Type of sequence clustering to be used in the phyloseq object, can be ESV, curatedESV, OTU, or curatedOTU.}

\item{assignment}{Taxonomic assignments from which to generate the taxa_table(). Can be "BLAST", "Idtaxa" or FALSE.}

\item{ClusterAssignment}{How clusters will recive taxonomic assignments. Can be "RepresentativeSequence" or between 0 and 1 to represent the proportion of reads that must share an assignment for an OTU to receive that assignment.}
}
\value{
A phyloseq object generated from the input SeqDataTable
}
\description{
This function takes the output of the bioinformatic pipeline into phyloseq.
}
\examples{
SeqDataTable2Phyloseq(SeqDataTablePath="23s_test_SeqDataTable.RDS", clustering="curatedOTU", assignment="BLAST", ClusterAssignment="RepresentativeSequence")
SeqDataTable2Phyloseq(SeqDataTablePath="~/Dropbox/BioinformaticPipeline_Env/Results/16s_multirun_test_SeqDataTable.RDS", clustering="ESV", assignment="Idtaxa")
}
