% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleMetaPackage-package.R
\docType{package}
\name{SimpleMetaPackage-package}
\alias{SimpleMetaPackage}
\alias{SimpleMetaPackage-package}
\title{SimpleMetaPackage: SimpleMetaPackage}
\description{
Trial bioinformatics package.
}
\author{
\strong{Maintainer}: Jake Williams \email{j.williams20@imperial.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
